Update README.md
Basically, the screen is divided into four quadrants to provide four discrete voices. The brighter the screen, the higher the pitch. This is done with the max/msp program from Cycling74 ( https://cycling74.com/ ). Many have used the program to provide video images to sound, but I am trying to reverse the process and create sound from video. This is similar to image sonification, used to create sound from data to help the vision-disabled to navigate their environment.